Alteração na view 'permissao_new'. :smile:

+++ pessoal/models.py
@@ -244,8 +244,8 @@ class Professor(models.Model):
         db_table = 'professor'
 
 class Reside(models.Model):
-    id = models.BigIntegerField(primary_key=True)
-    logradouro = models.ForeignKey('localizacao.Logradouro')    
+    id = models.AutoField(primary_key=True)
+    logradouro = models.ForeignKey('localizacao.Logradouro')
     numero = models.CharField(max_length=20, blank=True, null=True, verbose_name='Número')
     complemento = models.CharField(max_length=300, blank=True, null=True)
     cep = models.CharField(max_length=9, blank=True, null=True, verbose_name='CEP')
